inline util.js file
