+++ jsconfig.json
@@ -12,7 +12,7 @@
     "_types/**/*.d.ts",
     "*.js",
     "test/**/*"
-  ],
+, "experiment/util.js"  ],
   "exclude": [
     "node_modules",
     "test/lambdas/hello.js",
